requirements.txt
requests==2.28.1
urllib3==1.26.12
environs~=9.5.0
python-dotenv~=0.21.0